requirements.txt
Flask==2.3.4
ibm-watsonx-ai==1.1.20
Pillow==10.0.1
requests==2.32.0
